+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased] - 2021-08-21
+## [Unreleased] - 2021-03-29
 
 When next major version is released (3.x), these breaking changes will be applied:
  - remove deprecated `skip_lists` from all methods
@@ -11,13 +11,23 @@ When next major version is released (3.x), these breaking changes will be applie
 For such a version these issues will be resolved:
  - .copy() to actually render a full copy, instead of sharing data with the original
 
-## [2.2.17] - 2021-09-07
+## [2.2.19] - 2021-03-29
 
 ### Added
    - `.file_path` property, with matching `Path` value for `.filename`
 
 ### Fixes
    - resolve globals where `no_compound_keys` is set, and keys would be compound otherwise
+
+## [2.2.18] - 2022-02-04
+
+### Fixes
+   - avoid keys from dictionaries or files initialising a new Config to be overwritten with attribute values in object initialisation; this allows having keys with names like 'parameters'
+
+## [2.2.17] - 2022-02-03
+
+### Fixes
+   - fix formatting with globals, attempts at partial formatting a string with positional fields (like `'{:.2f} {some_global}'`) must not fail
 
 ## [2.2.16] - 2021-09-06
 
@@ -324,6 +334,8 @@ with cfg.direct:
   - First publicly released version after a major rewrite.
   
 [Unreleased]: /../../../
+[2.2.19]: /../../../tags/2.2.19
+[2.2.18]: /../../../tags/2.2.18
 [2.2.17]: /../../../tags/2.2.17
 [2.2.16]: /../../../tags/2.2.16
 [2.2.15]: /../../../tags/2.2.15